criando tabelas de pedidos e criando suas services e repositorys e ajustes nas trativas de erros

+++ src/services/IngredientsService.js
@@ -1,12 +1,22 @@
+const AppError = require("../utils/AppError");
+
 class IngredientsService {
     constructor(ingredientsRepository) {
         this.ingredientsRepository = ingredientsRepository;
     }
 
     async delete({ id }) {
-        const ingredientsDelete = await this.ingredientsRepository.deleteById({ id });
+        const ingredientsExists = await this.ingredientsRepository.findById({ id });
 
-        return ingredientsDelete;
+        if (!ingredientsExists) {
+            throw new AppError("Ingrediente não existe", 404);
+        }
+        try {
+            const ingredientsDelete = await this.ingredientsRepository.deleteById({ id });
+            return ingredientsDelete;
+        } catch (error) {
+            throw new AppError("Erro ao buscar detalhe", 400);
+        }
     }
 }
